added forbidden exception, readme file, crud users in group, moved inner parameter classes from controllers

+++ src/main/java/com/pet/chat/repository/GroupRepository.java
@@ -1,8 +1,8 @@
 package com.pet.chat.repository;
 
 import com.pet.chat.domain.Group;
-import com.pet.chat.domain.Message;
 import com.pet.chat.domain.User;
+import com.pet.chat.exception.NoPermissionException;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,13 +11,13 @@ import javax.persistence.PersistenceContext;
 import java.util.List;
 
 @Repository
+@Transactional
 public class GroupRepository {
     @PersistenceContext
     private EntityManager entityManager;
 
-    @Transactional
-    public void create(String name, long admin_id) {
-        User admin = entityManager.find(User.class, admin_id);
+    public void create(String name, long adminId) {
+        User admin = entityManager.find(User.class, adminId);
         Group group = new Group();
         group.setName(name);
         group.setAdmin(admin);
@@ -25,34 +25,31 @@ public class GroupRepository {
         entityManager.persist(group);
     }
 
-    @Transactional
-    public void remove(long group_id) {
-        Group group = find(group_id);
-        entityManager.remove(group);
-        entityManager.createQuery("delete from Message m where m.group_id=:group_id").setParameter("group_id", group_id).executeUpdate();
+    public void remove(long groupId, long adminId) {
+        Group group = find(groupId);
+        if (group.getId() == adminId) {
+            entityManager.remove(group);
+            entityManager.createQuery("delete from Message m where m.group_id=:group_id").setParameter("group_id", groupId).executeUpdate();
+        } else throw new NoPermissionException();
     }
 
-    @Transactional
     public List<Group> findAll() {
         return entityManager.createNativeQuery("select * from GROUPS", Group.class).getResultList();
     }
 
-    @Transactional
     public Group find(long id) {
         return entityManager.find(Group.class, id);
     }
 
-    @Transactional
-    public void addUser(long group_id, long user_id) {
-        Group group = find(group_id);
-        User user = entityManager.find(User.class, user_id);
+    public void addUser(long groupId, long userId) {
+        Group group = find(groupId);
+        User user = entityManager.find(User.class, userId);
         group.getUsers().add(user);
     }
 
-    @Transactional
-    public void removeUser(long group_id, long user_id) {
-        Group group = find(group_id);
-        User user = entityManager.find(User.class, user_id);
+    public void removeUser(long groupId, long userId) {
+        Group group = find(groupId);
+        User user = entityManager.find(User.class, userId);
         group.getUsers().remove(user);
     }
